optim: NativeArray, NativeInt and NativeString
don't embed Attr anymore,
and don't wrap array or string but directly declare them as aliases.

Gains are very modest however.

NativeUndefined also didn't bring much improvement and was let as is
(there are other leads for it). NativeMap offers ordered keys and
will stay like this for the foreseeing time